refactor: move export response-schema builders into serializers.py
The response-body schema builders (resource_responses,
build_resource_response_serializer, SyncErrorDetail) read straight off
the dynamic row serializers they document, so keep them in serializers.py
rather than a separate schema.py.
