--- apps/api/general/schema.py
@@ -1,47 +1,0 @@
-"""Builds the OpenAPI response-body schemas for the per-resource team-export sync endpoints. Each
-resource is documented with the shape of the same dynamic serializer the endpoint actually serves,
-so the docs can't drift from the payload. The per-resource view factory that attaches these schemas
-lives with the view in ``views.py``."""
-
-from drf_spectacular.utils import OpenApiResponse
-from rest_framework import serializers
-
-from apps.teams.export.manifest import ManifestEntry, entry_model
-
-from .serializers import build_resource_serializer, component_name
-
-
-class SyncErrorDetail(serializers.Serializer):
-    detail = serializers.CharField()
-
-
-def build_resource_response_serializer(entry: ManifestEntry) -> type[serializers.Serializer]:
-    """The paginated envelope the resource endpoint returns: cursor, has_more, and the page of rows."""
-    model = entry_model(entry.model)
-    item = build_resource_serializer(model)
-    return type(
-        f"{component_name(model)}List",
-        (serializers.Serializer,),
-        {
-            "cursor": serializers.CharField(
-                allow_null=True, help_text="Cursor for the next page; null once the resource is exhausted."
-            ),
-            "has_more": serializers.BooleanField(help_text="True while more rows remain beyond this page."),
-            "results": item(many=True),
-        },
-    )
-
-
-def resource_responses(entry: ManifestEntry) -> dict[int, type[serializers.Serializer] | OpenApiResponse]:
-    """Per-status responses for one resource: the 200 envelope, and a 409 (secret resources only)
-    when the team has no public key to seal against. No 404: routing prevents unknown resources from
-    reaching the view."""
-    responses: dict[int, type[serializers.Serializer] | OpenApiResponse] = {
-        200: build_resource_response_serializer(entry),
-    }
-    if entry.secret:
-        responses[409] = OpenApiResponse(
-            response=SyncErrorDetail,
-            description="Team has no registered public key; secret data cannot be sealed.",
-        )
-    return responses
