+++ apps/api/general/views.py
@@ -26,8 +26,13 @@ from apps.teams.export.manifest import (
 )
 from apps.teams.export.seal import load_public_key
 
-from .schema import resource_responses
-from .serializers import ManifestSerializer, build_resource_serializer, build_team_serializer, component_name
+from .serializers import (
+    ManifestSerializer,
+    build_resource_serializer,
+    build_team_serializer,
+    component_name,
+    resource_responses,
+)
 
 DEFAULT_LIMIT = 100
 MAX_LIMIT = 1000
